Proyecto Final Testing Etapa 4
This report summarizes the conclusions and results of the testing process carried out
throughout the eMercado project. The main objective of the testing was to evaluate
the quality, functionality, security and performance as general aspects of the system.
Throughout the various phases of the project, the testing team has worked diligently
to identify and address any issues that could impact the integrity and usability of the
system.